refactor: consolidate ErrorDetailSchema and authed_user into apps/core
Lifts the {detail: str} error schema and the request.user-narrowing
helper into shared homes so provenance (and future apps) can reuse them
without duplicating private copies.

- ErrorDetailSchema moves to apps/core/schemas.py; catalog's local copy
  and accounts' _ErrorSchema are deleted (no re-export shim).
- authed_user lands in new apps/core/api_helpers.py; citation and media
  drop their _authed_user twins.
